threat-intelligence-platform/requirements.txt
Django>=4.2
djangorestframework
celery
django-celery-beat
django-celery-results
redis
requests
python-dotenv
django-environ
elasticsearch
psycopg2-binary
gunicorn
uvicorn
